% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdist.R
\name{rdist}
\alias{rdist}
\title{simulation d'une loi discrete}
\usage{
rdist(x, p)
}
\arguments{
\item{x}{numeric vector representing the values of the random variable}

\item{p}{numeric vector representing the probabilities}
}
\description{
simulation d'une loi discrete
}
